issue-126, fft/brro compressor was renamed to atsc
