+++ atsc/src/compressor/polynomial.rs
