+++ atsc/src/optimizer/utils.rs
