+++ atsc/tests/integration_test.rs
@@ -102,7 +102,7 @@ fn compress_file_with_speed(speed: u8) {
 
 fn run_compressor(args: &[&str]) {
     // path to binary set by cargo: https://doc.rust-lang.org/cargo/reference/environment-variables.html#environment-variables-cargo-sets-for-crates
-    let command = std::env!("CARGO_BIN_EXE_brro-compressor");
+    let command = std::env!("CARGO_BIN_EXE_atsc");
 
     let status = std::process::Command::new(command)
         .args(args)
