+++ csv-compressor/Cargo.toml
@@ -2,12 +2,12 @@
 name = "csv-compressor"
 version = "0.1.0"
 edition = "2021"
-description = "Utilizes brro-compressor functionalities to compress CSV formatted data"
+description = "Utilizes ATSC functionalities to compress CSV formatted data"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-brro-compressor = { version = "0.5.0", path = "../brro-compressor" }
+atsc = { version = "0.5.0", path = "../atsc" }
 clap = { workspace = true, features = ["derive"] }
 csv = "1.3.0"
 serde = { version = "1.0.171", features = ["derive"] }
